+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### [0.1.7]
 
+Changed default config values
+
 ### [0.1.6]
 
 Fixed a bug where all actions would target the host instead of the desired target. It worked in local tests, but I realized the bug when trying the mod online with friends... Sorry!